Improve performance a bit

+++ algorithm/heavyhash-gate.h
@@ -8,7 +8,7 @@ struct xoshiro_state {
 };
 
 extern uint64_t le64dec(const void *pp);
-extern void generate_matrix(uint16_t matrix[64][64], struct xoshiro_state *state);
+extern void generate_matrix(uint32_t matrix[64][64], struct xoshiro_state *state);
 extern int heavyhash_test(unsigned char *pdata, const unsigned char *ptarget,
 			uint32_t nonce);
 extern void heavyhash_regenhash(struct work *work);
